% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{get_GSheet}
\alias{get_GSheet}
\title{Read the Google Sheet with index.html data}
\usage{
get_GSheet()
}
\value{
A data_frame of the index.html data
}
\description{
Read the Google Sheet with index.html data
}

