% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_boxes.R
\name{make_annot_boxes}
\alias{make_annot_boxes}
\title{Create bootstrap boxes for documents for Hypothes.is annotation}
\usage{
make_annot_boxes()
}
\description{
Create bootstrap boxes for documents for Hypothes.is annotation
}

