% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head.R
\name{make_head}
\alias{make_head}
\title{Create the <head> for ESC dev index.html}
\usage{
make_head()
}
\value{
An \code{htmltools tags$head} element
}
\description{
Edit this function if new styling components, names, etc. are needed
}
\examples{
\dontrun{
make_head()
}
}
\seealso{
\link{make_modals} \link{make_body}
}

