% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{make_modals}
\alias{make_modals}
\title{Create modals for index.html}
\usage{
make_modals()
}
\value{
A \code{htmltools div} with modal elements
}
\description{
Create modals for index.html
}
\examples{
mods <- make_modals()
}

