% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microapps.R
\name{make_muapp_boxes}
\alias{make_muapp_boxes}
\title{Create bootstrap boxes for Shiny microapps}
\usage{
make_muapp_boxes()
}
\value{
what is returned
}
\description{
Create bootstrap boxes for Shiny microapps
}

